YT-24309 Fix crashes related to missing TStatus::Repeat
Refactor code in a way that allows returning TStatus::Repeat early (when rebuilding section).
commit_hash:ffe00f7ba7ac0a52ff01bb5074c1112fd0c966b8
